+++ molsystem/__init__.py
@@ -24,14 +24,14 @@ import molsystem.elements  # noqa: F401
 from .system_db import SystemDB  # noqa: F401
 from .cell import Cell  # noqa: F401
 from .properties import standard_properties, add_properties_from_file  # noqa: F401
-from .openbabel import openbabel_version  # noqa: F401
+from .openbabel import openbabel_version, openbabel_citations  # noqa: F401
 from .align import RMSD  # noqa: F401
 
 try:
     from .openeye import openeye_version  # noqa: F401
 except ImportError:
     pass
-from .rdkit_ import rdkit_version  # noqa: F401
+from .rdkit_ import rdkit_version, rdkit_citations  # noqa: F401
 
 
 # Handle versioneer